pyproject.toml
[project]
name = "eegclip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "braindecode>=0.8.1",
    "hydra-core>=1.3.2",
    "jupyter>=1.1.1",
    "moabb>=1.2.0",
    "pytorch-lightning>=2.5.2",
    "transformers>=4.53.2",
    "wandb>=0.21.0",
]
